first integration test works
