+++ module #2. REST API Basics/rest-api-basic/build.gradle
@@ -43,6 +43,9 @@ allprojects {
         testImplementation group: 'org.mockito', name: 'mockito-junit-jupiter', version: '5.2.0'
         // https://mvnrepository.com/artifact/org.mockito/mockito-core
         testImplementation group: 'org.mockito', name: 'mockito-core', version: '5.2.0'
+        // https://mvnrepository.com/artifact/org.hamcrest/hamcrest
+        testImplementation group: 'org.hamcrest', name: 'hamcrest', version: '2.2'
+
 
 
         // https://mvnrepository.com/artifact/com.h2database/h2
@@ -50,9 +53,6 @@ allprojects {
         // https://mvnrepository.com/artifact/com.jayway.jsonpath/json-path
         implementation group: 'com.jayway.jsonpath', name: 'json-path', version: '2.8.0'
 
-
-        // https://mvnrepository.com/artifact/org.json/json
-        implementation group: 'org.json', name: 'json', version: '20230227'
 
         // https://mvnrepository.com/artifact/org.slf4j/slf4j-api
         implementation group: 'org.slf4j', name: 'slf4j-api', version: '2.0.7'
@@ -87,7 +87,7 @@ allprojects {
         // https://mvnrepository.com/artifact/org.apache.commons/commons-lang3
         implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.10'
         // https://mvnrepository.com/artifact/javax.servlet/javax.servlet-api
-        compileOnly group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
+        implementation group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
 
         // https://mvnrepository.com/artifact/com.bmuschko/gradle-tomcat-plugin
         implementation group: 'com.bmuschko', name: 'gradle-tomcat-plugin', version: '2.5'
