+++ module #2. REST API Basics/rest-api-basic/controller/src/main/java/com/epam/esm/configuration/AppConfig.java
