+++ module #2. REST API Basics/rest-api-basic/controller/src/main/resources/application-test.properties
