+++ module #2. REST API Basics/rest-api-basic/controller/src/test/java/TagIT.java
@@ -1,0 +1,50 @@
+import com.epam.esm.configuration.initializer.WebAppInitializer;
+import com.epam.esm.repository.configuration.datasource.impl.TestDataSourceConfig;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+import org.springframework.test.web.servlet.ResultMatcher;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import org.springframework.test.web.client.match.JsonPathRequestMatchers.*;
+import static org.hamcrest.Matchers.is;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = {WebAppInitializer.class, TestDataSourceConfig.class})
+@WebAppConfiguration()
+@ActiveProfiles("test")
+class TagIT {
+    @Autowired
+    private WebApplicationContext webApplicationContext;
+    private MockMvc mockMvc;
+
+    @BeforeEach
+    public void setup() throws Exception {
+        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.webApplicationContext).build();
+    }
+
+    @Test
+    void givenGreetURI_whenMockMVC_thenVerifyResponse() throws Exception {
+        MvcResult mvcResult = this.mockMvc.perform(get("/tags/find?name=tag5"))
+                .andDo(print()).andExpect(status().isOk())
+                .andExpect(jsonPath("$.name").value("tag5"))
+                .andReturn();
+
+        assertEquals("application/json",
+                mvcResult.getResponse().getContentType());
+    }
+}
